[no-issue] move package extension/bootstrap -> extension/
