+++ java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/sender/manager/MetricSendSessionManager.java
@@ -30,8 +30,8 @@ import java.util.concurrent.ThreadFactory;
 import java.util.logging.Logger;
 import org.traffichunter.javaagent.commons.status.AgentStatus;
 import org.traffichunter.javaagent.commons.util.AgentUtil;
-import org.traffichunter.javaagent.extension.bootstrap.AgentExecutableContext;
-import org.traffichunter.javaagent.extension.bootstrap.property.TrafficHunterAgentProperty;
+import org.traffichunter.javaagent.extension.AgentExecutableContext;
+import org.traffichunter.javaagent.extension.property.TrafficHunterAgentProperty;
 import org.traffichunter.javaagent.extension.metadata.AgentMetadata;
 import org.traffichunter.javaagent.extension.sender.websocket.AgentSystemMetricSender;
 import org.traffichunter.javaagent.extension.sender.websocket.AgentTransactionMetricSender;
