+++ java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/sender/websocket/AgentTransactionMetricSender.java
@@ -27,8 +27,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.traffichunter.javaagent.extension.metadata.AgentMetadata;
 import org.traffichunter.javaagent.extension.metadata.MetadataWrapper;
-import org.traffichunter.javaagent.extension.otel.TraceInfo;
-import org.traffichunter.javaagent.extension.otel.TraceQueue;
+import org.traffichunter.javaagent.extension.TraceInfo;
+import org.traffichunter.javaagent.extension.TraceQueue;
 import org.traffichunter.javaagent.extension.sender.MetricSender;
 import org.traffichunter.javaagent.websocket.MetricWebSocketClient;
 import org.traffichunter.javaagent.websocket.converter.SerializationByteArrayConverter.MetricType;
